tasklist4

+++ resources/views/tasks/show.blade.php
@@ -18,12 +18,13 @@
             <td>{{ $task->content }}</td>
         </tr>
     </table>
-    {{-- タスク内容編集ページへのリンク --}}
-    {!! link_to_route('tasks.edit', 'このタスクを編集', ['task' => $task->id], ['class' => 'btn btn-light']) !!}
-    
-    {{-- タスク内容削除フォーム --}}
-    {!! Form::model($task, ['route' => ['tasks.destroy', $task->id], 'method' => 'delete']) !!}
-        {!! Form::submit('削除', ['class' => 'btn btn-danger']) !!}
-    {!! Form::close() !!}
-
+    @if (Auth::id() == $task->user_id)
+        {{-- タスク内容編集ページへのリンク --}}
+        {!! link_to_route('tasks.edit', 'このタスクを編集', ['task' => $task->id], ['class' => 'btn btn-light']) !!}
+        
+        {{-- タスク内容削除フォーム --}}
+        {!! Form::model($task, ['route' => ['tasks.destroy', $task->id], 'method' => 'delete']) !!}
+            {!! Form::submit('削除', ['class' => 'btn btn-danger']) !!}
+        {!! Form::close() !!}
+    @endif
 @endsection
